backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
PyJWT==2.11.0
bcrypt==5.0.0
psycopg[binary]==3.2.13
redis==6.4.0
pytest==9.0.2
httpx==0.28.1
